+++ payment/tests/test_models.py
@@ -1,3 +1,2 @@
 # test credit_no == 16 only
 # test ccv is either 3 or 4 only
-# test credit_no cannot be the same
